QuickTask/analytics_service/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pymongo>=4.5.0
python-dotenv>=1.0.0
pydantic>=2.0.0
python-multipart>=0.0.6
